+++ Back-end/Squirrel/Requests/User/LoginRequest.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Squirrel.Models
+namespace Squirrel.Requests.User
 {
-    public class LoginModel
+    public class LoginRequest
     {
 
         [DataType(DataType.EmailAddress)]
